Refactor `wallet` cmd
