--- src/commands/wallet/address.rs
@@ -1,8 +1,0 @@
-use crate::config::ZKSyncConfig;
-use eyre::ContextCompat;
-
-pub(crate) async fn run(cfg: ZKSyncConfig) -> eyre::Result<()> {
-    let wallet_config = cfg.wallet.context("Wallet config missing")?;
-    println!("Wallet address: {:?}", wallet_config.address);
-    Ok(())
-}
