--- src/commands/wallet/withdraw.rs
@@ -1,1 +1,0 @@
-
